Concatenated values of word net suggestions
The values that are suggested from wordnet are here concatenated. this
is done for further indexing and retrieval operations in GATE ANNIC
platform. That is for saving in searchable lucene datastore, and
performing querying on the corpus.
In addition soundex value for each word is added. This is done for
searching at homophone confusion